Add deduce-all (and deduce-all-with)
This seeks to add some type of applictative functor -ish functionality,
but in a way that is usable in this library.

Where >>= takes one value and pushes it through all the given functions,
deduce-all takes one fn and many deducer-values, calls f with them all,
and gathers the contexts back together. Maybe not exactly <*>, but...

  * MultiDeducer protocol: deduce-all calls apply-all, in this protocol
  * deduce-all-with; like deducer->>=, it wraps and unwraps using a fn
  * Update docstrings
  * Test coverage for deduce-all. Well, OK, for deduce-all-with.

TODO
 * Simplify some test cases to be less convoluted and easier to understand
 * transformers (covered by compose?)
 * curried fn, maybe